Fixed small Error and Changed Quotes to more appropriate ones
The other ones were imagined in a hurry

+++ src/main/kotlin/com/silverwyrm/quote/control/QuoteDao.kt
@@ -17,9 +17,9 @@ open class QuoteDao {
         return e.resultList
     }
 
-    open fun count(): Int {
-        val e = em.createNamedQuery("quote.count", Int::class.java)
-        return e.singleResult
+    open fun count(): Long {
+        val e = em.createNamedQuery("quote.count", java.lang.Long::class.java)
+        return e.singleResult as Long
     }
 
     @Transactional
